Moves the file of pomodoro_lib target in a shared folder
