+++ src/shared/Configuration.cpp
