+++ src/shared/Configuration.hpp
