+++ src/shared/ConfigurationFile.cpp
