+++ src/shared/ConfigurationFile.hpp
