+++ src/shared/utility/FilesDictionary.cpp
