+++ src/shared/utility/FilesDictionary.hpp
