+++ src/shared/utility/StringsDictionary.hpp
